chore: added signature based minting

+++ src/core/DotHypeResolver.sol
@@ -18,31 +18,39 @@ import "../../lib/ens-contracts/contracts/resolvers/ResolverBase.sol";
  * @dev Resolver contract for DotHype domains
  * Implements ENS compatible resolution interfaces
  */
-contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolver, ITextResolver, IContentHashResolver, IReverseResolver {
+contract DotHypeResolver is
+    Ownable,
+    ResolverBase,
+    IAddrResolver,
+    IAddressResolver,
+    ITextResolver,
+    IContentHashResolver,
+    IReverseResolver
+{
     // Custom errors
     error NotAuthorized(address caller, bytes32 node);
     error InvalidNode(bytes32 node);
     error RecordNotFound(bytes32 node, string recordType);
     error DomainExpired(bytes32 node, uint256 expiry);
-    
+
     // Reference to the DotHype registry (immutable for gas savings and security)
     IDotHypeRegistry public immutable registry;
-    
+
     // Event for version changes
     event VersionChanged(bytes32 indexed node, uint256 newVersion);
-    
+
     // Mapping from node to record version
     mapping(bytes32 => uint256) private _recordVersions;
-    
+
     // Storage for domain records with versioning
     mapping(bytes32 => mapping(uint256 => address)) private _addresses;
     mapping(bytes32 => mapping(uint256 => mapping(uint256 => bytes))) private _coinAddresses;
     mapping(bytes32 => mapping(uint256 => mapping(string => string))) private _textRecords;
     mapping(bytes32 => mapping(uint256 => bytes)) private _contentHashes;
-    
+
     // Simple reverse resolution mapping
     mapping(address => bytes32) private _reverseRecords; // address => domain node
-    
+
     /**
      * @dev Modifier to ensure domain is active (not expired)
      * @param node The namehash of the domain
@@ -57,7 +65,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
             revert InvalidNode(node);
         }
     }
-    
+
     /**
      * @dev Constructor
      * @param _owner Initial owner of the contract
@@ -66,7 +74,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
     constructor(address _owner, address _registry) Ownable(_owner) {
         registry = IDotHypeRegistry(_registry);
     }
-    
+
     /**
      * @dev Returns the current version of records for a node
      * @param node The namehash of the domain
@@ -75,7 +83,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
     function recordVersion(bytes32 node) public view returns (uint256) {
         return _recordVersions[node];
     }
-    
+
     /**
      * @dev Increments the version for a domain's records, effectively clearing all records
      * @param node The namehash of the domain
@@ -84,7 +92,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
         _recordVersions[node]++;
         emit VersionChanged(node, _recordVersions[node]);
     }
-    
+
     /**
      * @dev Checks if a domain is active (not expired)
      * @param node The namehash of the domain
@@ -97,7 +105,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
             return false;
         }
     }
-    
+
     /**
      * @dev Returns the address associated with a domain
      * @param node The namehash of the domain
@@ -109,7 +117,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
         }
         return payable(_addresses[node][_recordVersions[node]]);
     }
-    
+
     /**
      * @dev Sets the address associated with a domain
      * @param node The namehash of the domain
@@ -120,32 +128,32 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
         _addresses[node][version] = a;
         emit AddrChanged(node, a);
     }
-    
+
     /**
      * @dev Returns the address associated with a domain for a specific coin type
      * @param node The namehash of the domain
      * @param coinType The coin type as per SLIP-0044
      * @return The associated address
      */
-    function addr(bytes32 node, uint coinType) public view virtual override returns (bytes memory) {
+    function addr(bytes32 node, uint256 coinType) public view virtual override returns (bytes memory) {
         if (!isActive(node)) {
             return "";
         }
         return _coinAddresses[node][_recordVersions[node]][coinType];
     }
-    
+
     /**
      * @dev Sets the address associated with a domain for a specific coin type
      * @param node The namehash of the domain
      * @param coinType The coin type as per SLIP-0044
      * @param a The address to set
      */
-    function setAddr(bytes32 node, uint coinType, bytes memory a) public virtual authorised(node) onlyActive(node) {
+    function setAddr(bytes32 node, uint256 coinType, bytes memory a) public virtual authorised(node) onlyActive(node) {
         uint256 version = _recordVersions[node];
         _coinAddresses[node][version][coinType] = a;
         emit AddressChanged(node, coinType, a);
     }
-    
+
     /**
      * @dev Returns the text data associated with a domain
      * @param node The namehash of the domain
@@ -158,19 +166,24 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
         }
         return _textRecords[node][_recordVersions[node]][key];
     }
-    
+
     /**
      * @dev Sets the text data associated with a domain
      * @param node The namehash of the domain
      * @param key The text record key
      * @param value The text record value
      */
-    function setText(bytes32 node, string calldata key, string calldata value) public virtual authorised(node) onlyActive(node) {
+    function setText(bytes32 node, string calldata key, string calldata value)
+        public
+        virtual
+        authorised(node)
+        onlyActive(node)
+    {
         uint256 version = _recordVersions[node];
         _textRecords[node][version][key] = value;
         emit TextChanged(node, key, key, value);
     }
-    
+
     /**
      * @dev Returns the content hash associated with a domain
      * @param node The namehash of the domain
@@ -182,7 +195,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
         }
         return _contentHashes[node][_recordVersions[node]];
     }
-    
+
     /**
      * @dev Sets the content hash associated with a domain
      * @param node The namehash of the domain
@@ -193,7 +206,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
         _contentHashes[node][version] = hash;
         emit ContenthashChanged(node, hash);
     }
-    
+
     /**
      * @dev Gets the domain node that an address points to (reverse resolution)
      * @param addr The address to lookup
@@ -202,7 +215,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
     function getNode(address addr) public view override returns (bytes32) {
         return _reverseRecords[addr];
     }
-    
+
     /**
      * @dev Sets the domain node for an address (reverse resolution)
      * Can only be called by the owner of the address
@@ -216,15 +229,15 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
         } catch {
             revert InvalidNode(node);
         }
-        
+
         // Sender must own the domain they're pointing to
         require(msg.sender == domainOwner, "Not domain owner");
-        
+
         // Set the reverse record
         _reverseRecords[msg.sender] = node;
         emit ReverseResolutionSet(msg.sender, node);
     }
-    
+
     /**
      * @dev Clears the reverse record for the sender's address
      */
@@ -232,7 +245,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
         _reverseRecords[msg.sender] = bytes32(0);
         emit ReverseResolutionCleared(msg.sender);
     }
-    
+
     /**
      * @dev Gets the domain name for an address through reverse resolution
      * @param addr The address to lookup
@@ -240,12 +253,12 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
      */
     function reverseLookup(address addr) public view override returns (string memory) {
         bytes32 node = _reverseRecords[addr];
-        
+
         // Return empty if no reverse record or domain is expired
         if (node == bytes32(0) || !isActive(node)) {
             return "";
         }
-        
+
         // Return the domain name
         try registry.tokenIdToName(uint256(node)) returns (string memory domainName) {
             return string(abi.encodePacked(domainName, ".hype"));
@@ -253,7 +266,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
             return "";
         }
     }
-    
+
     /**
      * @dev Gets the domain name for an address through reverse resolution (alias for reverseLookup)
      * Also checks if the address matches the address record in the resolver
@@ -262,18 +275,18 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
      */
     function getName(address addr) public view override returns (string memory) {
         bytes32 node = _reverseRecords[addr];
-        
+
         // Return empty if no reverse record or domain is expired
         if (node == bytes32(0) || !isActive(node)) {
             return "";
         }
-        
+
         // Check that the address record matches the lookup address
         address registeredAddress = _addresses[node][_recordVersions[node]];
         if (registeredAddress != addr) {
             return ""; // Address doesn't match the resolver record
         }
-        
+
         // Return the domain name
         try registry.tokenIdToName(uint256(node)) returns (string memory domainName) {
             return string(abi.encodePacked(domainName, ".hype"));
@@ -281,7 +294,7 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
             return "";
         }
     }
-    
+
     /**
      * @dev Gets a specific value for an address through reverse resolution
      * Also checks if the address matches the address record in the resolver
@@ -291,22 +304,22 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
      */
     function getValue(address addr, string calldata key) public view override returns (string memory) {
         bytes32 node = _reverseRecords[addr];
-        
+
         // Return empty if no reverse record or domain is expired
         if (node == bytes32(0) || !isActive(node)) {
             return "";
         }
-        
+
         // Check that the address record matches the lookup address
         address registeredAddress = _addresses[node][_recordVersions[node]];
         if (registeredAddress != addr) {
             return ""; // Address doesn't match the resolver record
         }
-        
+
         // Return the text record value
         return _textRecords[node][_recordVersions[node]][key];
     }
-    
+
     /**
      * @dev Checks if an address has a reverse record
      * Also checks if the address matches the address record in the resolver
@@ -315,17 +328,17 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
      */
     function hasRecord(address addr) public view override returns (bool) {
         bytes32 node = _reverseRecords[addr];
-        
+
         // Return false if no reverse record or domain is expired
         if (node == bytes32(0) || !isActive(node)) {
             return false;
         }
-        
+
         // Check that the address record matches the lookup address
         address registeredAddress = _addresses[node][_recordVersions[node]];
         return registeredAddress == addr;
     }
-    
+
     /**
      * @dev Checks if the resolver supports a specific interface
      * This overrides the ResolverBase implementation
@@ -333,15 +346,11 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
      * @return True if the interface is supported
      */
     function supportsInterface(bytes4 interfaceID) public view virtual override returns (bool) {
-        return
-            interfaceID == type(IAddrResolver).interfaceId ||
-            interfaceID == type(IAddressResolver).interfaceId ||
-            interfaceID == type(ITextResolver).interfaceId ||
-            interfaceID == type(IContentHashResolver).interfaceId ||
-            interfaceID == type(IReverseResolver).interfaceId ||
-            super.supportsInterface(interfaceID);
+        return interfaceID == type(IAddrResolver).interfaceId || interfaceID == type(IAddressResolver).interfaceId
+            || interfaceID == type(ITextResolver).interfaceId || interfaceID == type(IContentHashResolver).interfaceId
+            || interfaceID == type(IReverseResolver).interfaceId || super.supportsInterface(interfaceID);
     }
-    
+
     /**
      * @dev Overrides the authorised function in ResolverBase
      * Checks if the sender is authorized to update the record
@@ -355,4 +364,4 @@ contract DotHypeResolver is Ownable, ResolverBase, IAddrResolver, IAddressResolv
             return false;
         }
     }
-} 
+}
